requirements.txt
# Core Streamlit framework
streamlit==1.29.0

# Database and ORM
sqlalchemy==2.0.23

# Data analysis and visualization (essential for analytics)
pandas==2.1.4
plotly==5.17.0

# Security and authentication
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0
PyJWT==2.8.0

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Email services
sendgrid==6.10.0
boto3==1.34.0  # AWS SES support

# Web framework for webhooks
flask==3.0.0

# Basic utilities
requests==2.31.0
email-validator==2.1.0

# HTML processing and validation
beautifulsoup4==4.12.2
lxml==4.9.3

# File handling and data export
openpyxl==3.1.2

# Monitoring and logging
sentry-sdk==1.38.0
